# swagger.yaml
openapi: 3.0.3
info:
  title: md-to-pdf Document Engine API
  description: |
    Moteur de documents complet : conversion Markdown/HTML/Template vers PDF,
    preview PNG, fusion, watermark et protection par mot de passe.

    ## Comportement conditionnel des réponses

    La plupart des endpoints de génération PDF acceptent les champs optionnels
    `client_id` et `pdf_name`. Selon leur présence :

    | `client_id` + `pdf_name` | Réponse |
    |--------------------------|---------|
    | **fournis** | Le PDF est sauvegardé sur le serveur et un JSON `{ "download_url": "/download/…" }` est retourné |
    | **absents** | Le fichier PDF binaire est retourné directement dans le body |
  version: 1.0.0
  contact:
    name: AI SmartTalk
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Développement local
  - url: https://pdf.aismarttalk.tech
    description: Production
...
paths:
...
  /api/convert:
    post:
      tags: [Convert]
      summary: Markdown → PDF
      description: |
        Convertit du Markdown en PDF via pandoc + moteur PDF (weasyprint par défaut).

        Supporte les tags CENSOR (`{{CENSOR}}`, `<CENSOR>`, etc.) qui sont remplacés
        par une image floutée pour le contenu premium.

        `header_html` est prioritaire sur `header_template` (inline > fichier serveur).
      operationId: convert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConvertRequest"
            example:
              markdown: "# Mon Document\n\nContenu en **Markdown**."
              css: "h1 { color: navy; }"
              engine: weasyprint
              options:
                paper_size: a4
                orientation: portrait
                page_numbers: true
              header_template: "header.html"
              client_id: "client-123"
              pdf_name: "rapport-2024"
      responses:
        "200":
          description: PDF généré avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConvertResponse"
              example:
                download_url: "/download/client-123/rapport-2024.pdf"
            application/pdf:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ───────────────────────────────────────────────
  # Render  (Template Tera + JSON → PDF)
  # ───────────────────────────────────────────────
...
  /api/preview:
    post:
      tags: [Preview]
      summary: Génération de preview PNG (première page)
      description: |
        Génère un PDF à partir de l'un des trois modes d'entrée, puis retourne
        une image PNG de la première page (150 DPI).

        **Modes d'entrée** (mutuellement exclusifs, par ordre de priorité) :
        1. `markdown` — conversion Markdown → PDF → PNG
        2. `template` + `data` — rendu Tera → PDF → PNG
        3. `html` — HTML → PDF → PNG

        Retourne toujours `image/png`, jamais de JSON.
      operationId: preview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreviewRequest"
            examples:
              markdown:
                summary: Preview d'un document Markdown
                value:
                  markdown: "# Aperçu\n\nTexte de test."
              html:
                summary: Preview d'un HTML
                value:
                  html: "<html><body><h1>Preview HTML</h1></body></html>"
              template:
                summary: Preview d'un template Tera
                value:
                  template: "<h1>{{ title }}</h1>"
                  data:
                    title: "Mon titre"
      responses:
        "200":
          description: Image PNG de la première page
          content:
            image/png:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ───────────────────────────────────────────────
  # Merge
  # ───────────────────────────────────────────────
...
components:
...
  schemas:
    PdfEngine:
      type: string
      enum: [weasyprint, wkhtmltopdf, pdflatex]
      default: weasyprint
      description: Moteur PDF utilisé par pandoc
    PaperSize:
      type: string
      enum: [a4, a3, letter]
      default: a4
    Orientation:
      type: string
      enum: [portrait, landscape]
      default: portrait
    Margins:
      type: object
      description: Marges de la page (valeurs CSS, ex. "2cm", "1in", "20mm")
      properties:
        top:
          type: string
          example: "2cm"
        bottom:
          type: string
          example: "2cm"
        left:
          type: string
          example: "2cm"
        right:
          type: string
          example: "2cm"
    PdfOptions:
      type: object
      description: Options de rendu PDF communes à tous les endpoints de génération
      properties:
        paper_size:
          $ref: "#/components/schemas/PaperSize"
        orientation:
          $ref: "#/components/schemas/Orientation"
        margins:
          $ref: "#/components/schemas/Margins"
        page_numbers:
          type: boolean
          default: false
          description: Afficher les numéros de page en bas de chaque page
        page_number_format:
          type: string
          default: "counter(page)"
          description: Format CSS du numéro de page (ex. `counter(page) "/" counter(pages)`)
        toc:
          type: boolean
          default: false
          description: Générer une table des matières (pandoc uniquement)
        toc_depth:
          type: integer
          minimum: 1
          maximum: 6
          description: Profondeur de la table des matières
        watermark:
          type: string
          description: Texte de watermark léger appliqué via CSS
    ConvertRequest:
      type: object
      required: [markdown]
      properties:
        markdown:
          type: string
          description: Contenu Markdown à convertir. Supporte les tags CENSOR.
        css:
          type: string
          description: CSS additionnel (ajouté après le CSS par défaut)
        engine:
          $ref: "#/components/schemas/PdfEngine"
        options:
          $ref: "#/components/schemas/PdfOptions"
        header_html:
          type: string
          description: HTML inline pour le header (prioritaire sur header_template)
        footer_html:
          type: string
          description: HTML inline pour le footer (prioritaire sur footer_template)
        header_template:
          type: string
          description: Nom du fichier template header (dans le dossier templates/)
        footer_template:
          type: string
          description: Nom du fichier template footer (dans le dossier templates/)
        client_id:
          type: string
          description: Identifiant client pour sauvegarder le PDF sur le serveur
        pdf_name:
          type: string
          description: Nom du fichier PDF (l'extension .pdf est ajoutée automatiquement)
...
    PreviewRequest:
      type: object
      description: |
        Fournir **un seul** des modes d'entrée :
        - `markdown` pour une conversion Markdown
        - `template` + `data` pour un rendu Tera
        - `html` pour du HTML direct
      properties:
        markdown:
          type: string
          description: Contenu Markdown (mode Markdown)
        html:
          type: string
          description: Contenu HTML (mode HTML)
        template:
          type: string
          description: Template Tera (mode Template, requiert `data`)
        data:
          type: object
          additionalProperties: true
          description: Données JSON pour le template (mode Template)
        css:
          type: string
        engine:
          $ref: "#/components/schemas/PdfEngine"
        options:
          $ref: "#/components/schemas/PdfOptions"
        header_html:
          type: string
        footer_html:
          type: string
        header_template:
          type: string
        footer_template:
          type: string
...
    ConvertResponse:
      type: object
      properties:
        download_url:
          type: string
          description: URL relative pour télécharger le PDF généré
          example: "/download/client-123/rapport-2024.pdf"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Type d'erreur
          example: "Bad request"
        details:
          type: string
          description: Détails / message d'erreur
          example: "At least 2 PDFs are required for merging"

  # ─── Responses réutilisables ──────────────────
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Bad request"
            details: "\"data\" must be a JSON object"
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Not found"
            details: "PDF not found: /download/client-123/missing.pdf"
    InternalError:
      description: Erreur interne du serveur
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Pandoc conversion failed"
            details: "pandoc: ... (stderr output)"